+++ lib/ardecy/harden.rb
@@ -17,26 +17,7 @@ module Ardecy
     def self.modules(args)
       puts
       title 'Kernel Modules'
-
-      Modules::Blacklist::Dccp.new(args).x
-      Modules::Blacklist::Sctp.new(args).x
-      Modules::Blacklist::Rds.new(args).x
-      Modules::Blacklist::Tipc.new(args).x
-      Modules::Blacklist::NHdlc.new(args).x
-      Modules::Blacklist::Ax25.new(args).x
-      Modules::Blacklist::Netrom.new(args).x
-      Modules::Blacklist::X25.new(args).x
-      Modules::Blacklist::Rose.new(args).x
-      Modules::Blacklist::Decnet.new(args).x
-      Modules::Blacklist::Econet.new(args).x
-      Modules::Blacklist::Af802154.new(args).x
-      Modules::Blacklist::Ipx.new(args).x
-      Modules::Blacklist::Appletalk.new(args).x
-      Modules::Blacklist::Psnap.new(args).x
-      Modules::Blacklist::P8023.new(args).x
-      Modules::Blacklist::P8022.new(args).x
-      Modules::Blacklist::Can.new(args).x
-      Modules::Blacklist::Atm.new(args).x
+      Modules::Blacklist.exec(args)
       return unless args[:fix]
 
       if Dir.exist? '/etc/modprobe.d/'
@@ -62,27 +43,7 @@ module Ardecy
 
     def self.sysctl_kernel(args)
       title 'Kernel Hardening'
-
-      Sysctl::Kernel::KPointer.new(args).x
-      Sysctl::Kernel::Dmesg.new(args).x
-      Sysctl::Kernel::Printk.new(args).x
-      Sysctl::Kernel::BpfDisabled.new(args).x
-      Sysctl::Kernel::BpfJitHarden.new(args).x
-      Sysctl::Kernel::LdiskAutoload.new(args).x
-      Sysctl::Kernel::UserFaultFd.new(args).x
-      Sysctl::Kernel::KExecLoadDisabled.new(args).x
-      Sysctl::Kernel::SysRQ.new(args).x
-      Sysctl::Kernel::UsernsClone.new(args).x
-      Sysctl::Kernel::MaxUserNameSpace.new(args).x
-      Sysctl::Kernel::PerfEventParanoid.new(args).x
-      Sysctl::Kernel::YamaPtrace.new(args).x
-      Sysctl::Kernel::VmMmapRndBits.new(args).x
-      Sysctl::Kernel::VmMmapRndCompatBits.new(args).x
-      Sysctl::Kernel::FsProtectedSymlinks.new(args).x
-      Sysctl::Kernel::FsProtectedHardlinks.new(args).x
-      Sysctl::Kernel::FsProtectedFifos.new(args).x
-      Sysctl::Kernel::FsProtectedRegular.new(args).x
-
+      Sysctl::Kernel.exec(args)
       return unless args[:fix]
 
       if Dir.exist? '/etc/sysctl.d/'
@@ -95,30 +56,7 @@ module Ardecy
 
     def self.sysctl_network(args)
       title 'Network Hardening'
-
-      Sysctl::Network::TcpSynCookie.new(args).x
-      Sysctl::Network::RFC1337.new(args).x
-      Sysctl::Network::AllRpFilter.new(args).x
-      Sysctl::Network::DefaultRpFilter.new(args).x
-      Sysctl::Network::AllAcceptRedirects.new(args).x
-      Sysctl::Network::DefaultAcceptRedirects.new(args).x
-      Sysctl::Network::AllSecureRedirects.new(args).x
-      Sysctl::Network::DefaultSecureRedirects.new(args).x
-      Sysctl::Network::Ipv6AllAcceptRedirects.new(args).x
-      Sysctl::Network::Ipv6DefaultAcceptRedirects.new(args).x
-      Sysctl::Network::AllSendRedirects.new(args).x
-      Sysctl::Network::DefaultSendRedirects.new(args).x
-      Sysctl::Network::IcmpEchoIgnoreAll.new(args).x
-      Sysctl::Network::AllAcceptSourceRoute.new(args).x
-      Sysctl::Network::DefaultAcceptSourceRoute.new(args).x
-      Sysctl::Network::Ipv6AllAcceptSourceRoute.new(args).x
-      Sysctl::Network::Ipv6DefaultAcceptSourceRoute.new(args).x
-      Sysctl::Network::Ipv6ConfAllAcceptRa.new(args).x
-      Sysctl::Network::Ipv6ConfDefaultAcceptRa.new(args).x
-      Sysctl::Network::TcpSack.new(args).x
-      Sysctl::Network::TcpDSack.new(args).x
-      Sysctl::Network::TcpFack.new(args).x
-
+      Sysctl::Network.exec(args)
       return unless args[:fix]
 
       if Dir.exist? '/etc/sysctl.d/'